backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
firebase-admin==6.6.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
msgpack==1.1.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.8.0
pyparsing==3.2.3
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.3
rpds-py==0.27.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.7.0
xlsxwriter==3.2.5
gunicorn==22.0.0
sentry-sdk==1.45.0
